+++ core/hadoop-common/src/main/scala/com/webank/wedatasphere/linkis/hadoop/common/utils/HDFSUtils.scala
@@ -1,38 +1,20 @@
-/*
- * Copyright 2019 WeBank
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.webank.wedatasphere.linkis.common.utils
+package com.webank.wedatasphere.linkis.hadoop.common.utils
 
 import java.io.File
 import java.nio.file.Paths
 import java.security.PrivilegedExceptionAction
 
-import com.webank.wedatasphere.linkis.common.conf.CommonVars
-import com.webank.wedatasphere.linkis.common.conf.Configuration.{HADOOP_ROOT_USER, KERBEROS_ENABLE, KEYTAB_FILE, kEYTAB_HOST,KEYTAB_HOST_ENABLED}
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, Path}
+import com.webank.wedatasphere.linkis.hadoop.common.conf.HadoopConf._
 import org.apache.hadoop.security.UserGroupInformation
-
+import com.webank.wedatasphere.linkis.common.conf.Configuration.hadoopConfDir
 /**
   * Created by enjoyyin on 2019/5/27.
   */
 object HDFSUtils {
 
-  val hadoopHome = CommonVars("HADOOP_HOME", CommonVars("hadoop.home", "").getValue).getValue
-  val hadoopConfDir = CommonVars("HADOOP_CONF_DIR", CommonVars("hadoop.config.dir", "").getValue).getValue
+
 
   def getConfiguration(user: String): Configuration = getConfiguration(user, hadoopConfDir)
